+++ tests/unit/connector/test_opensearch_output.py
@@ -136,7 +136,7 @@ class TestOpenSearchOutput(BaseOutputTestCase):
         assert failed_document == expected
 
     @mock.patch(
-        "opensearchpy.helpers.bulk",
+        "opensearchpy.helpers.parallel_bulk",
         side_effect=search.SerializationError,
     )
     def test_write_to_search_context_calls_handle_serialization_error_if_serialization_error(
@@ -148,7 +148,7 @@ class TestOpenSearchOutput(BaseOutputTestCase):
         self.object._handle_serialization_error.assert_called()
 
     @mock.patch(
-        "opensearchpy.helpers.bulk",
+        "opensearchpy.helpers.parallel_bulk",
         side_effect=search.ConnectionError,
     )
     def test_write_to_search_context_calls_handle_connection_error_if_connection_error(self, _):
@@ -158,7 +158,7 @@ class TestOpenSearchOutput(BaseOutputTestCase):
         self.object._handle_connection_error.assert_called()
 
     @mock.patch(
-        "opensearchpy.helpers.bulk",
+        "opensearchpy.helpers.parallel_bulk",
         side_effect=helpers.BulkIndexError,
     )
     def test_write_to_search_context_calls_handle_bulk_index_error_if_bulk_index_error(self, _):
@@ -167,7 +167,7 @@ class TestOpenSearchOutput(BaseOutputTestCase):
         self.object._write_to_search_context({"dummy": "event"})
         self.object._handle_bulk_index_error.assert_called()
 
-    @mock.patch("opensearchpy.helpers.bulk")
+    @mock.patch("opensearchpy.helpers.parallel_bulk")
     def test__handle_bulk_index_error_calls_bulk(self, fake_bulk):
         mock_bulk_index_error = mock.MagicMock()
         mock_bulk_index_error.errors = [
@@ -181,7 +181,7 @@ class TestOpenSearchOutput(BaseOutputTestCase):
         self.object._handle_bulk_index_error(mock_bulk_index_error)
         fake_bulk.assert_called()
 
-    @mock.patch("opensearchpy.helpers.bulk")
+    @mock.patch("opensearchpy.helpers.parallel_bulk")
     def test_handle_bulk_index_error_calls_bulk_with_error_documents(self, fake_bulk):
         mock_bulk_index_error = mock.MagicMock()
         mock_bulk_index_error.errors = [